[#47011] Replace OAuth::TableCell by ViewComponent
https://community.openproject.org/work_packages/47011
